+++ docs-web/.vitepress/theme/index.ts
@@ -55,6 +55,7 @@ import EvInput from '../../../packages/evoke-ui/src/components/input/index.vue'
 import EvTextarea from '../../../packages/evoke-ui/src/components/textarea/index.vue'
 import EvSelect from '../../../packages/evoke-ui/src/components/select/index.vue'
 import EvField from '../../../packages/evoke-ui/src/components/field/index.vue'
+import EvSlider from '../../../packages/evoke-ui/src/components/slider/index.vue'
 import EvPullRefresh from '../../../packages/evoke-ui/src/components/pull-refresh/index.vue'
 import EvLoadMore from '../../../packages/evoke-ui/src/components/load-more/index.vue'
 import EvActionSheet from '../../../packages/evoke-ui/src/components/action-sheet/index.vue'
@@ -86,7 +87,7 @@ export default {
       EvTimeline, EvComparisonTable, EvCta, EvNewsletter, EvLogoCloud,
       EvVideo, EvAudio, EvContactForm, EvCarousel, EvArticleCard, EvProfileCard,
       EvBorderBeam, EvExecCard, EvArticle, EvImageWall, EvImagePreview, EvWaterfall, EvModal,
-      EvMarquee, EvInput, EvTextarea, EvSelect, EvField,
+      EvMarquee, EvInput, EvTextarea, EvSelect, EvField, EvSlider,
       EvPullRefresh, EvLoadMore, EvActionSheet, EvTabbar, EvTabbarItem, EvNavBar,
       DemoBlock, MobileStage,
     }